api/requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
